pyproject.toml
[project]
name = "neon-city-chess"
version = "1.0.0"
description = "Cinematic two-player chess with a neon cyberpunk aesthetic"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Boyd Roberts" }]
keywords = ["chess", "opengl", "cyberpunk", "game"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment :: Board Games",
]
dependencies = [
    "python-chess>=1.999",
    "moderngl>=5.10.0",
    "glfw>=2.7.0",
    "numpy>=1.24.0",
    "pyrr>=0.10.3",
]

[project.scripts]
neon-chess = "main:main"

[project.urls]
Repository = "https://github.com/coleyrockin/Chess-Game"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.bandit]
skips = ["B311"]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
